backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
alembic==1.12.1

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.0.0

# Don't specify pydantic-core - it's automatically installed with pydantic